Split accuracy test for general reduction ops
