+++ tests/test_prod.py
@@ -1,0 +1,55 @@
+import pytest
+import torch
+
+import flag_gems
+
+from . import accuracy_utils as utils
+from . import conftest as cfg
+
+if cfg.QUICK_MODE:
+    FLOAT_DTYPES = [torch.float32]
+    DIM_LIST = [0]
+    KEEPDIM = [True]
+else:
+    FLOAT_DTYPES = utils.FLOAT_DTYPES
+    DIM_LIST = [0, 1]
+    KEEPDIM = [True, False]
+
+
+@pytest.mark.prod
+@pytest.mark.parametrize("shape", utils.REDUCTION_SHAPES)
+@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
+def test_prod(shape, dtype):
+    if flag_gems.vendor_name == "kunlunxin":
+        torch.manual_seed(0)
+        torch.cuda.manual_seed_all(0)
+
+    inp = torch.randn(shape, dtype=dtype, device=flag_gems.device)
+    ref_inp = utils.to_reference(inp, True)
+
+    ref_out = torch.prod(ref_inp)
+    with flag_gems.use_gems():
+        res_out = torch.prod(inp)
+
+    utils.gems_assert_close(res_out, ref_out, dtype)
+
+
+# TODO: failed at (200, 40999, 3), while successed at this shape in mean_dim
+@pytest.mark.prod_dim_int
+@pytest.mark.parametrize("shape", utils.REDUCTION_SMALL_SHAPES)
+@pytest.mark.parametrize("keepdim", KEEPDIM)
+@pytest.mark.parametrize("dim", DIM_LIST)
+@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
+def test_prod_dim_int(shape, dim, keepdim, dtype):
+    if flag_gems.vendor_name == "kunlunxin":
+        torch.manual_seed(0)
+        torch.cuda.manual_seed_all(0)
+
+    inp = torch.randn(shape, dtype=dtype, device=flag_gems.device)
+    ref_inp = utils.to_reference(inp, True)
+
+    ref_out = torch.prod(ref_inp, dim=dim, keepdim=keepdim)
+    with flag_gems.use_gems():
+        res_out = torch.prod(inp, dim=dim, keepdim=keepdim)
+
+    utils.gems_assert_close(res_out, ref_out, dtype)
